third_party/blink/web_tests/fast/css-generated-content/pseudo-animation-expected.txt

Animations on :before and :after pseudo elements should run

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS div.offsetWidth is 52
PASS div.offsetWidth is within 1 of 20
PASS computedTop is within 1 of 170
PASS div.offsetWidth is within 1 of 12
PASS computedTop is within 1 of 200
PASS div.offsetWidth is 52
PASS div.offsetWidth is within 1 of 20
PASS computedTop is within 1 of 170
PASS div.offsetWidth is within 1 of 12
PASS computedTop is within 1 of 200
PASS successfullyParsed is true

TEST COMPLETE

 
